Storage: Separate DMFile (#9042)
ref pingcap/tiflash#6233
