+++ dbms/src/Flash/Disaggregated/MockS3LockClient.h
@@ -43,7 +43,7 @@ public:
         // If the data file exist and no delmark exist, then create a lock file on `data_file_key`
         auto view = S3FilenameView::fromKey(data_file_key);
         auto object_key
-            = view.isDMFile() ? fmt::format("{}/{}", data_file_key, DM::DMFile::metav2FileName()) : data_file_key;
+            = view.isDMFile() ? fmt::format("{}/{}", data_file_key, DM::DMFileMetaV2::metaFileName()) : data_file_key;
         if (!objectExists(*s3_client, object_key))
         {
             return {false, ""};
