+++ dbms/src/Flash/Disaggregated/tests/gtest_s3_lock_service.cpp
@@ -68,7 +68,7 @@ public:
                 DMFileOID{.store_id = store_id, .table_id = physical_table_id, .file_id = dm_file_id});
             DB::S3::uploadEmptyFile(
                 *s3_client,
-                fmt::format("{}/{}", data_filename.toFullKey(), DM::DMFile::metav2FileName()));
+                fmt::format("{}/{}", data_filename.toFullKey(), DM::DMFileMetaV2::metaFileName()));
             ++dm_file_id;
         }
     }
