+++ dbms/src/Storages/DeltaMerge/Delta/DeltaValueSpace.cpp
@@ -181,7 +181,7 @@ std::vector<ColumnFilePtrT> CloneColumnFilesHelper<ColumnFilePtrT>::clone(
                 file_id,
                 /* page_id= */ new_page_id,
                 file_parent_path,
-                DMFile::ReadMetaMode::all(),
+                DMFileMeta::ReadMode::all(),
                 dm_context.keyspace_id);
 
             auto new_column_file = f->cloneWith(dm_context, new_file, target_range);
