+++ dbms/src/Storages/DeltaMerge/Remote/DataStore/DataStoreMock.cpp
@@ -35,7 +35,7 @@ static std::tuple<String, UInt64> parseDMFilePath(const String & path)
     return std::tuple<String, UInt64>{parent_path, file_id};
 }
 
-DMFilePtr MockPreparedDMFileToken::restore(DMFile::ReadMetaMode read_mode)
+DMFilePtr MockPreparedDMFileToken::restore(DMFileMeta::ReadMode read_mode)
 {
     auto [parent_path, file_id] = parseDMFilePath(path);
     return DMFile::restore(
