+++ dbms/src/Storages/DeltaMerge/Remote/DataStore/DataStoreS3.h
@@ -78,7 +78,7 @@ public:
 
     ~S3PreparedDMFileToken() override = default;
 
-    DMFilePtr restore(DMFile::ReadMetaMode read_mode) override;
+    DMFilePtr restore(DMFileMeta::ReadMode read_mode) override;
 };
 
 } // namespace DB::DM::Remote
