+++ dbms/src/Storages/DeltaMerge/Segment.cpp
@@ -1374,7 +1374,7 @@ SegmentPtr Segment::dangerouslyReplaceDataFromCheckpoint(
         data_file->fileId(),
         new_page_id,
         data_file->parentPath(),
-        DMFile::ReadMetaMode::all(),
+        DMFileMeta::ReadMode::all(),
         dm_context.keyspace_id);
     wbs.data.putRefPage(new_page_id, data_file->pageId());
 
@@ -1415,7 +1415,7 @@ SegmentPtr Segment::dangerouslyReplaceDataFromCheckpoint(
             auto remote_data_store = dm_context.global_context.getSharedContextDisagg()->remote_data_store;
             RUNTIME_CHECK(remote_data_store != nullptr);
             auto prepared = remote_data_store->prepareDMFile(file_oid, new_data_page_id);
-            auto dmfile = prepared->restore(DMFile::ReadMetaMode::all());
+            auto dmfile = prepared->restore(DMFileMeta::ReadMode::all());
             auto new_column_file = b->cloneWith(dm_context, dmfile, rowkey_range);
             new_column_file_persisteds.push_back(new_column_file);
         }
@@ -1813,14 +1813,14 @@ Segment::prepareSplitLogical( //
             file_id,
             /* page_id= */ my_dmfile_page_id,
             file_parent_path,
-            DMFile::ReadMetaMode::all(),
+            DMFileMeta::ReadMode::all(),
             dm_context.keyspace_id);
         auto other_dmfile = DMFile::restore(
             dm_context.global_context.getFileProvider(),
             file_id,
             /* page_id= */ other_dmfile_page_id,
             file_parent_path,
-            DMFile::ReadMetaMode::all(),
+            DMFileMeta::ReadMode::all(),
             dm_context.keyspace_id);
         my_stable_files.push_back(my_dmfile);
         other_stable_files.push_back(other_dmfile);
