+++ dbms/src/Storages/DeltaMerge/tests/gtest_dm_delta_merge_store_fast_add_peer.cpp
@@ -330,7 +330,7 @@ try
                     file_id.id,
                     file_id.id,
                     delegator.getDTFilePath(file_id.id),
-                    DMFile::ReadMetaMode::all());
+                    DMFileMeta::ReadMode::all());
                 remote_store->putDMFile(
                     dm_file,
                     S3::DMFileOID{
