+++ dbms/src/Storages/DeltaMerge/tests/gtest_segment_reader.cpp
@@ -199,7 +199,7 @@ try
     const auto & dmfiles = store->id_to_segment.begin()->second->getStable()->getDMFiles();
     ASSERT_EQ(dmfiles.size(), 1);
     auto dmfile = dmfiles.front();
-    auto readable_path = DMFile::getPathByStatus(dmfile->parentPath(), dmfile->fileId(), DMFile::Status::READABLE);
+    auto readable_path = getPathByStatus(dmfile->parentPath(), dmfile->fileId(), DMFileStatus::READABLE);
     ASSERT_EQ(dmfile->path(), readable_path);
     ASSERT_EQ(DMFileReaderPool::instance().get(readable_path), nullptr);
