+++ dbms/src/Storages/DeltaMerge/tests/gtest_segment_test_basic.cpp
@@ -526,7 +526,7 @@ void SegmentTestBasic::ingestDTFileIntoDelta(
             file_id,
             ref_id,
             parent_path,
-            DMFile::ReadMetaMode::all());
+            DMFileMeta::ReadMode::all());
         wbs.writeLogAndData();
         ASSERT_TRUE(segment->ingestDataToDelta(
             *dm_context,
@@ -585,7 +585,7 @@ void SegmentTestBasic::ingestDTFileByReplace(
             file_id,
             ref_id,
             parent_path,
-            DMFile::ReadMetaMode::all());
+            DMFileMeta::ReadMode::all());
         wbs.writeLogAndData();
 
         auto apply_result = segment->ingestDataForTest(*dm_context, ref_file, clear);
