+++ dbms/src/Storages/Page/V3/Universal/tests/gtest_checkpoint.cpp
@@ -1093,7 +1093,7 @@ try
         S3::uploadEmptyFile(*s3_client, ingest_from_data_file.toFullKey());
         S3::uploadEmptyFile(
             *s3_client,
-            fmt::format("{}/{}", ingest_from_dtfile.toFullKey(), DM::DMFile::metav2FileName()));
+            fmt::format("{}/{}", ingest_from_dtfile.toFullKey(), DM::DMFileMetaV2::metaFileName()));
 
         UniversalWriteBatch batch;
